# adventures/src/main/java/com/ynov/adventures/dto/AventurierCreateDTO.java
package com.ynov.adventures.dto;

import com.ynov.adventures.domain.Classe;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class AventurierCreateDTO {

    @NotBlank(message = "Le nom est obligatoire")
    @Size(min = 2, max = 100, message = "Le nom doit faire entre 2 et 100 caractères")
    private String nom;

    @Size(max = 1000, message = "La description ne peut pas dépasser 1000 caractères")
    private String description;

    @NotNull(message = "Les caractéristiques sont obligatoires")
    @Valid
    private CaracteristiquesDTO caracteristiques;

    @NotNull(message = "La classe est obligatoire")
    private Classe classe;
}